Add explicit cast

+++ src/ruuvi_endpoint_5.c
@@ -110,7 +110,7 @@ static void re_5_encode_temperature (uint8_t * const buffer, const re_5_data_t *
     if (!isnan (temperature))
     {
         re_clip (&temperature, RE_5_TEMP_MIN, RE_5_TEMP_MAX);
-        int16_t rounded_temperature = roundf (temperature * RE_5_TEMP_RATIO);
+        int16_t rounded_temperature = (int16_t) roundf (temperature * RE_5_TEMP_RATIO);
         // Type cast adds 2^16 to a negative signed value, not changing bits.
         coded_temperature = (uint16_t) rounded_temperature;
     }
